Move TestUtil to test package
